Update reindex count tests
Here we ensure that we are counting the same objects before and after
the reindex call so that the test results are valid. The issue why ACR
is now being indexed is that it's pulled into a session with each object
automatically. Before there were no ACLs created but reindex still
indexed their entires but that means that the previous test did not
actually compare same numbers, but they just happen to be the same.

This commit also changes the count of objects from 5 to 3 to still
leave a plural count of individual objects to get the correct test
results and to improve test speed.