Fix response processing
We couldn't take all information from request in some cases. If we
update ticket in Issue Tracker, we send not all information (we can't
send ccs for example cause we can erase some information). So, as we
wouldn't send people, we should populate it with information from
Issue Tracker(in response we can see all information about ticket
no doubt what was in request) or just skip this update. I've chosen
to  populate it.