GGRC-8324 Add possibility to import csv files with semicolon delimiter
Now it's possible to import csv's with comma or semicolon as delimiters.
In case of other delimiters, error is raised and user would get pop-up
message with error.
Add integration tests and 2 test csv files - one with semicolons, which is imported
without errors, and another with $-sign as delimiter, which cannot be imported - are also added.