Fix workflow unarchivation
According to the requirements, workflow unarchivation via API should be prohibited.
However, it is possible to do that by sending PUT request with a body containing
field "recurrences" set to true.

To fix the bug I have added validate_recurrences method to the Workflow model.
The method will raise ValueError if field "recurrences" in the payload is true
leading to the response with status 400(Bad request) being returned by the API.

TCM test_unarchive_workflow_returns_status_400 was added to validate the fix.
The TCM verifies the result of the attempt to unarchive is the response with status 400.

I have added WorkflowArchivedFactory as a single step to create
an archived workflow. This factory is used in test_unarchive_workflow_returns_status_400.

Resolves: GGRC-8285